Move old IAP tests to an extension
